safu/log: Add logging macros for errors with return values
Adds the macros safuLogErrValue() and safuLogErrErrnoValue() which
contain the usual error string combined with the returned value
from the failed function to provide more information during errors.

This will also handle the case with 'errno' where the returned value is not -1,
where errno is not guaranteed to be set properly.
This is done since it does not make sense to bloat our code
(and our tests as we have to test this too for coverage) with additional
conditional code that will likely never be executed. We instead provide
the details in the log that are necessary to figure out if errno is
valid or not.